Rename table, ignore empty messages
